ml/requirements.txt
# python = 3.10

nba_api==1.5.2
pytorch_forecasting==1.1.1
PyYAML
tqdm
numpy

# pytorch forecasting
sphinx >3.2
nbsphinx
pandoc
docutils
pydata-sphinx-theme
lightning >=2.0.0
cloudpickle
torch >=2.0,!=2.0.1
optuna >=3.1.0
optuna-integration
scipy
pandas >=1.3
scikit-learn >1.2
matplotlib
statsmodels
ipython
nbconvert >=6.3.0
recommonmark >=0.7.1
pytorch-optimizer >=2.5.1
fastapi >0.80
cpflows